+++ core/cmd/app.go
@@ -459,21 +459,27 @@ func NewApp(client *Client) *cli.App {
 							Action: client.ExportETHKey,
 						},
 						{
-							Name:   "reset",
-							Usage:  "Manually set the next nonce for a key. This should not be necessary during normal operation. USE WITH CAUTION: Setting this incorrectly can break your node",
-							Action: client.ResetETHKey,
+							Name:   "chain",
+							Usage:  "Update an EVM key for the given chain",
+							Action: client.UpdateChainEVMKey,
 							Flags: []cli.Flag{
 								cli.StringFlag{
-									Name:  "address",
-									Usage: "address of the key for which to set the nonce",
+									Name:     "address",
+									Usage:    "address of the key",
+									Required: true,
 								},
 								cli.StringFlag{
-									Name:  "evmChainID",
-									Usage: "key nonce will be reset for this chain ID only",
+									Name:     "evmChainID",
+									Usage:    "chain ID of the key",
+									Required: true,
 								},
 								cli.Uint64Flag{
-									Name:  "nextNonce",
-									Usage: "the next nonce in the sequence",
+									Name:  "setNextNonce",
+									Usage: "manually set the next nonce for the key on the given chain. This should not be necessary during normal operation. USE WITH CAUTION: Setting this incorrectly can break your node",
+								},
+								cli.BoolFlag{
+									Name:  "setEnabled",
+									Usage: "enable/disable the key for the given chain",
 								},
 							},
 						},